server/requirements.txt
# Core Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

# AI Agent Framework (Updated to latest compatible versions with checkpoint support)
langgraph==0.2.58
langchain==0.3.10
langchain-core==0.3.25
langchain-openai==0.2.12
langchain-google-genai==2.0.6

# Data Processing & Analysis
pandas==2.2.3
numpy==1.26.4
openpyxl>=3.1.0
xlrd==2.0.1
python-docx==1.1.2

# Vector Store & Embeddings (Updated)
chromadb==0.5.23
sentence-transformers==3.3.1

# Data Validation & Serialization (Updated)
pydantic==2.10.3
pydantic-settings==2.6.1

# Database (SQLite & ORM)
sqlalchemy==2.0.36
alembic==1.14.0

# Background Tasks & Queue
celery==5.4.0
redis==5.2.0

# Image Processing (for OCR) - 
pillow==11.0.0
opencv-python==4.10.0.84
easyocr==1.7.2

# Audio Processing (for transcription)
pydub==0.25.1
SpeechRecognition==3.14.3
openai-whisper==20231117

# HTTP & API
httpx==0.28.1
requests==2.31.0
aiofiles==24.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sse-starlette==2.1.3

# Environment & Configuration
python-dotenv==1.0.1
python-decouple==3.8

# Logging & Monitoring
structlog==24.4.0
sentry-sdk[fastapi]==2.19.0

# Testing & Development
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
faker==33.1.0

# Data Science & Statistics
scipy==1.14.1
scikit-learn==1.6.0
statsmodels==0.14.4
seaborn==0.13.2
matplotlib==3.10.0

# Utilities (Removed uuid - it's built-in)
python-magic==0.4.27
chardet==5.2.0
pytz==2024.2

# Development Tools
black==24.10.0
flake8==7.1.1
isort==5.13.2
mypy==1.13.0

# Documentation
mkdocs==1.6.1
mkdocs-material==9.5.47
